mapping file refactor
