+++ oda_data/indicators/dac1/dac1_functions.py
@@ -1,9 +1,11 @@
 import pandas as pd
 
-from oda_data.logger import logger
-
 
 def official_oda(data: pd.DataFrame) -> pd.DataFrame:
+    """Filter the data to only include Official Development Assistance (ODA).
+    Pre-2018, ODA is defined as aidtype_code 1010.
+    Post-2018, ODA is defined as aidtype_code 11010.
+    """
     query = "(year < 2018 & aidtype_code == 1010) | (year >2018 & aidtype_code ==11010)"
 
     data = data.query(query)
